All daos made that pertain to students; add student post request now working

+++ hospitalschool/src/main/java/com/unc/hospitalschool/model/County.java
@@ -16,7 +16,7 @@ public class County {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name = "cid", nullable = false)
-	private Long cid;
+	private int cid;
 	
 	@Column(name = "county", nullable = false)
 	private String county;
